requirements.txt
click
logging
requests
datetime
aiohttp
backoff
psycopg2
python-dotenv